requirements.txt
# requirements for signald
attrs
-e git+https://github.com/mobilecoinofficial/full-service.git@ef1ab817fbfb1c9a2e44de6a2af774c2e84079e8#egg=mobilecoin&subdirectory=cli

base58==2.1.0
django==3.0.4
psycopg2==2.9.1
uwsgi==2.0.19.1
pyzt
googlemaps
protobuf==3.17.3

git+https://gitlab.com/mobilecoin/pysignald.git@b2f247df1edd2400608d4e003f4235496e7c630e#egg=pysignald
django-extensions==3.1.3
aenum==3.1.0
django-phonenumber-field[phonenumberslite]==2.1.0

# Make manage.py shell use an iPython session
ipython
factory_boy==3.2.0

